refactor(fmt): split tests/cli.rs into themed submodules
Pure move: all 2545 lines of `tests/cli.rs` land verbatim in
`tests/cli/main.rs` plus ten themed submodules (`common`, `stdin`,
`write_check`, `style`, `config`, `delegation`, `native`, `daemon`,
`sidecar`, `tailwind`). The `tests/cli/main.rs` layout keeps this a
single test binary — one link, not eleven. Only `use` lines and
visibility (`pub(crate)` on the shared helpers) were adjusted; the same
64 tests run before and after.
